requirements.txt
typing-extensions>=4.8.0
requests==2.31.0
psutil==5.9.8
websockets==15.0.1
PyYAML==6.0.2